feat: semantic highlighting for terminal output
Add a semantic highlight engine (crates/highlight) that classifies
plain-text terminal output — paths, numbers, MAC addresses, URLs,
commands, options, log levels, errors, dates, strings — and applies
distinct colors via a configurable theme (crates/ui/assets/highlight/default.json).

- New SemanticHighlightingMode (Auto/On/Off) in LayoutConfig, persisted
  to terminal.json, with a dropdown control in Settings → Terminal → Layout
- Scanner with shell-profile-aware prompt detection (Unix/Cmd/PowerShell/Dumb),
  backward-extending path probe, and structural matchers for numbers, IPs,
  MACs, URLs, strings, errors, log levels, and datetime patterns
- Integration in the terminal render pipeline (SemanticOverlay) with
  per-cell color override and theme-driven palette
- Test script (scripts/test_highlight.sh) and design doc
  (docs/terminal-semantic-highlighting.md)

+++ crates/ui/src/views/terminal/url/mask.rs
@@ -19,7 +19,7 @@ pub fn url_column_mask(line_cells: &[&IndexedCell]) -> Vec<bool> {
     let max_col = line_cells
         .iter()
         .filter(|ic| !ic.cell.flags.contains(Flags::WIDE_CHAR_SPACER))
-        .map(|ic| ic.point.column.0 as usize)
+        .map(|ic| ic.point.column.0)
         .max()
         .unwrap_or(0);
     let n = max_col + 1;
@@ -31,7 +31,7 @@ pub fn url_column_mask(line_cells: &[&IndexedCell]) -> Vec<bool> {
         if ic.cell.flags.contains(Flags::WIDE_CHAR_SPACER) {
             continue;
         }
-        let col = ic.point.column.0 as usize;
+        let col = ic.point.column.0;
         if col >= n {
             continue;
         }
@@ -59,7 +59,10 @@ pub fn url_column_mask(line_cells: &[&IndexedCell]) -> Vec<bool> {
             if i + plen > n {
                 continue;
             }
-            let matches = prefix.iter().zip(&chars[i..i + plen]).all(|(a, b)| *a == *b);
+            let matches = prefix
+                .iter()
+                .zip(&chars[i..i + plen])
+                .all(|(a, b)| *a == *b);
             if !matches {
                 continue;
             }
@@ -151,8 +154,10 @@ pub fn url_masks_wrapped(
                 if i + plen > n {
                     continue;
                 }
-                let matches =
-                    prefix.iter().zip(&chars[i..i + plen]).all(|(a, b)| *a == *b);
+                let matches = prefix
+                    .iter()
+                    .zip(&chars[i..i + plen])
+                    .all(|(a, b)| *a == *b);
                 if !matches {
                     continue;
                 }
